ERP/requirements.txt
# === CORE APP ===
streamlit==1.51.0
pandas==2.3.3
numpy==2.0.2

# === SUPABASE (VERSIÓN ESTABLE + COMPLETA) ===
supabase==2.4.0   # SDK unificado
postgrest==0.13.0 # compatible
gotrue==2.4.3
realtime==2.0.2
httpx==0.28.1

# === BASE DE DATOS / SQL ===
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# === OPENAI / IA ===
openai==1.51.0

# === VISUALIZACIÓN ===
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0

# === UTILIDADES ===
python-dotenv==1.0.1
requests==2.32.3
